version2/mainproject/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
beautifulsoup4==4.12.3
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
distro==1.9.0
Django==5.1
django-cors-headers==4.4.0
django-ninja==1.3.0
firebase-admin==6.5.0
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.17.2
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.5
grpcio-status==1.65.5
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
jiter==0.5.0
msgpack==1.0.8
openai==1.42.0
proto-plus==1.24.0
protobuf==5.27.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
pyparsing==3.1.2
requests==2.32.3
rsa==4.9
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
